Add files via upload
The file is a continuation of the data cleansing and descriptive analysis conducted earlier. It contains the steps taken in order to deal with the issue of an imbalanced dataset (Includes several techniques such as; using different performance metrics to analyze a model, re-sampling methods like SMOTE, Tomeklinks, Cluster Centroid, Using a tree based algorithm, etc.). In addition, it also entails some of the commonly used feature selection and reduction methods such as Variance Threshold, RFE, PCA, etc to get the most influential features. Finally, some hyper parameter tuning was conducted to enhance the predictive power of the models.